+++ platform/src/codat_platform/models/shared/datatypewritewebhookpayload.py
@@ -171,7 +171,7 @@ class DataTypeWriteWebhookPayload(BaseModel):
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
-            val = serialized.get(k)
+            val = serialized.get(k, serialized.get(n))
             is_nullable_and_explicitly_set = (
                 k in nullable_fields
                 and (self.__pydantic_fields_set__.intersection({n}))  # pylint: disable=no-member
@@ -186,3 +186,9 @@ class DataTypeWriteWebhookPayload(BaseModel):
                     m[k] = val
 
         return m
+
+
+try:
+    DataTypeWriteWebhookPayload.model_rebuild()
+except NameError:
+    pass